Merge main branch to database branch

+++ app/Http/Controllers/EmailVerificationController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\BaseController;
 use App\Mail\SendCodeEmailVerification;
 use App\Models\EmailVerification;
 use App\Models\User;
@@ -10,14 +11,15 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Mail;
 use LDAP\Result;
+use Illuminate\Support\Facades\Validator;
 
 use function PHPUnit\Framework\isNull;
 
-class EmailVerificationController extends Controller
+class EmailVerificationController extends BaseController
 {
     public function userCheckCode(Request $request) : JsonResponse
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'email' => 'required|email|exists:email_verifications,email',
             'code' => 'required|string|exists:email_verifications,code',
         ],[
@@ -25,25 +27,27 @@ class EmailVerificationController extends Controller
             'code.exists' => 'Code is not valid',
         ]);
 
-        // find the code
-        $emailverification = EmailVerification::query()->firstWhere('email', $request['email']);
-
-        if ($request['code'] != $emailverification['code'])
+        if($validator->fails())
         {
-            return response()->json([
-                'status' => 'failure',
-                'message' => 'Code is not valid',
-            ]);
+            return $this->sendError($validator->errors());
+        }
+
+        // find the code
+        $email_verification = EmailVerification::query()->firstWhere('email', $request['email']);
+
+        if ($request['code'] != $email_verification['code'])
+        {
+            return $this->sendError(['error' => trans('Code is not valid')]);
         }
 
         // check if it is not expired : the time is one hour
-        if ($emailverification['created_at'] > now()->addHour()) {
-            $emailverification->delete();
-            return response()->json(['status' => 'failure', 'message' => trans('email.code_is_expire')], 422);
+        if ($email_verification['created_at'] > now()->addHour()) {
+            $email_verification->delete();
+            return $this->sendError(['error' => trans('password.code_is_expire')]);
         }
 
         // find user's email
-        $user = User::query()->where('email', $emailverification['email']);
+        $user = User::query()->where('email', $email_verification['email']);
 
         // update user email_verified
         $user->update([
@@ -51,20 +55,26 @@ class EmailVerificationController extends Controller
         ]);
 
         // delete current code
-        $emailverification->delete();
+        $email_verification->delete();
 
         return response()->json([
             'status' => 'success',
-            'code' => $emailverification['code'],
+            'code' => $email_verification['code'],
             'message' => trans('email.code_is_valid'),
-        ]);
+        ], 200);
     }
 
     public function userResendCode(Request $request) : JsonResponse
     {
-        $data = $request->validate([
+        $data = $request->all();
+        $validator = Validator::make($data, [
             'email' => 'required|email|exists:users,email',
         ]);
+
+        if($validator->fails())
+        {
+            return $this->sendError($validator->errors());
+        }
 
         // Delete all old code that user send before
         EmailVerification::query()->where('email', $request['email'])->delete();
